rename service tests to service tester
