+++ service-tester/run_tests.sh
